ISS-461 add nullables without strictNullChecks
This commit only adds type information of nullables for when we add
strictNullChecks later. Some accesses of nullables without explicit
check are suppressed with the "!" operator - this is only recommended
for existing (and tested) code, or unit tests (for convenience).

The compiler output (i.e. runtime behaviour) is unchanged.

To finally activate strictNullChecks, some minor behaviour changes
will be necessary, which will be in another PR.

Signed-off-by: Benjamin Große <benjamin@midokura.com>