de/serialization test of steps
It might be desirable to serialize the server step 1. For example, in a
typical webserver, the connection is stateless and there is some kind of
session mechanism that has some storage layer.

It is possible to save username+salt+verifier in that session storage,
so that step 1 can be recreated.

For convenience, this patch adds a `serialize` module that can stringify
server and client steps and rehydrate with the `deserialize` complement.

I don't think client session steps should be serialized, however it is
also supported.